Use ftp to get files, better realignment code
Turns our that GE scanners run an ftp server on the private network.
So, I wrote the rtclient class to use ftp to interact with the scanner.
This seems faster and more reliable than the old NFS-based code. It also
means less external set-up (that is, if your scanner already runs ftp.)
I've also started to make the volumizer more robust (needs more work)
and switched to use the nipy fmri_realign_4d module, which is more
appropriate for within-subject motion correction. However, that code
is still clunky and slow due to their api, which is awkward given what
we are trying to do (align on the fly rather than after-the-fact).
We should maybe extend their api to allow what we want to be done
more efficiently.