Seems to be working now. Just need auto-update
The basic functionality is there, and seems solid. However, with this
version, the user needs to hit refresh to update the graph. The data are
curently pulled in through AJAX, so it should be easy to get it to
auto-update every second or so. (Just need to understand the rickshaw
api a little better.) The code could also use a little cleaning. And we
might consider going to a thin webapp framework, such as cherrypy or
webapp2 to tidy things up even more.