Rework SeriesFinder and add DicomFinder
This undoes some of the changes in previous versions, which ended
up being too clever for their own good. The queue objects themselves
now have to be created externally to the finder objects, which have
been renamed appropriately. It's hard to tell exactly what the
problem was, as it was causing errors to get returned from the FTP
server, but it was fixed by making the queue objects an argument to
the finder constructor rather than creating them internally.

Unfortunately, this does not fix the issue with the unit tests,
which are still failing with a "connection refused error" from
the FTP server, but only when trying to access it from within
one of the finder objects.