Rework how the DicomFinder updates the series it's operating on
The previous version was broken in that all of the dicom finding was
under a test for a non-empty series queue, meaning that the finder
would grab the first couple of dicoms and then hang around until a
new series got put on the queue. That's fixed here. This change also
flushes the dicom_files set each time a new series arrives so that
it does not get overly large.