Fix ordering of files
There is a hack in the code that uses the file order in addition to the
timestamp, because the latter only has minute-resolution. The problem is
that, although DICOM files are named sequentially, they are not zero-padded,
and so the sorted order that we get from the FTP dir() method is not really
the order that we want.

This adds an alphanumeric sort. We should now get dicoms off the scanner
in the right order, which will facilitate the volumizer.