Add initial code for motion analyzer
Currently the main loop of the analyzer thread is just a schematic
of what needs to happen.

This also very well may be changed in the future when I hear back
from the nipy list about the best way to do a quick and (not too)
dirty image-to-image registration.

Currently this uses the realign4d stuff, and not optimally, but
it's quite slow and we're gonna be right up against the TR limit
for real-time stuff.