Streamline motion analyzer code
The major changes here are removal of the code that registered each
volume to the previous volume to compute the relative RMS deviation.
Instead, that deviation can be computed from just the previous
transformation matrix (to the reference volume).

This also simplifies a few tests that were trick to keep passing,
and implements (but does not fully use) a couple of tricks from
Alexis Roche that makes the nipy registration code usable.