Use a different approach to determine needed slice numbers
Previous approach assumed that the first slice into the queue was
the first slice from the acquisition. That is not always true.
This approach computes the needed slices externally to the slices it
actually gets.

To do so it explicitly checks the exam/series/acq number of each
incoming slice and determines if a new acquisition has started.
This may cause problems if dicoms are written out of order with
respect to acquisitions. Hopefully that will not happen.