Allow SeriesFinder to skip early series
This should make it easier to start the realtime motion tracker
midway through the scan.

Currently this doesn't seem to fully work. While it picks up the
specified scan correctly, it doesn't flip over to the next scan
when it finishes. The SeriesFinder does find a new series after
the subsequent one starts, but then a bunch of Dicoms get dumped
into the DicomFinder and they never get built into a full image.

Clearly, something is misaligned at some point.

The other issue is that the number of series that the SeriesFinder
sees does not seem to correspond to the number of series that the
scanner thinks it has run. This might be due to files like fieldmaps
that maybe do not end up on the FTP server in real time? This also
could be related to the issue above, I'm not sure.