Updated interface and added utilities
The interface now accepts an option to not use series finder
Instead we can either get the lastest series, or predict the next
series if we're on the scanner and need to trigger a new scan.

The utilities allow us to trigger the scan, and also contain alphanum_key
as it is now used in the masker as well as the client.

Some small changes to client.